Fix: Error in 'Parse_with_lexer'
Error: If the lexer cannot complete a token immediately before the end
of input, then the error is not reported as a syntax error in the lexer.
It is reported as a syntax error in the parser which can be missleading.

Fix: The lexer is required to recognized an end token at the end of the
input stream which consumes the end of input. The module
'Parse_with_lexer' recognizes the end token of the lexer by calling the
function 'has_consumed_end' in the state 'has_succeeded'.

	modified:   src/parse/parse_with_lexer.ml
	modified:   src/parse/test_json.ml
	modified:   src/parse/test_token_yaml.ml