Fix fmlib_js: Element.focus, Element.blur
A javascript method which can accept zero or more arguments must not be
called with the ocaml '()' of type unit. It has to be called with
'Js.null'.

Reason: In Javascript each method has an arity. If in an actual call
fewer arguments are provided, then the method is called with 'null' for
the missing arguments. If more arguments than the arity requires then
the additional arguments are ignored.

	modified:   src/js/dom.ml