Redesign of subscriptions
In the state we have now two dirty flags. One for the dom and one for
the subscriptions. When the state changes, both dirty flags are set.

The function 'update_dom' sets the dirty flag for the dom to 'false'.
The function 'update_subscriptions' sets the dirty flag for the
subscriptions to 'false'.

Both functions are now much simpler than before.

	modified:   src/browser/browser.ml
	modified:   src/browser/command.ml
	modified:   src/browser/subscriptions.ml
	modified:   src/examples/browser/webapp.ml